+++ cli/cli/helpers/metrics_client_factory/metrics_client_factory.go
@@ -3,12 +3,12 @@ package metrics_client_factory
 import (
 	"github.com/kurtosis-tech/kurtosis/cli/cli/defaults"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/helpers/do_nothing_metrics_client_callback"
-	"github.com/kurtosis-tech/kurtosis/cli/cli/helpers/logrus_logger_converter"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/helpers/metrics_user_id_store"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/kurtosis_cluster_setting"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/kurtosis_config"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/kurtosis_config/resolved_config"
 	"github.com/kurtosis-tech/kurtosis/kurtosis_version"
+	"github.com/kurtosis-tech/metrics-library/golang/lib/analytics_logger"
 	metrics_client "github.com/kurtosis-tech/metrics-library/golang/lib/client"
 	"github.com/kurtosis-tech/metrics-library/golang/lib/source"
 	"github.com/kurtosis-tech/stacktrace"
@@ -51,7 +51,7 @@ func GetMetricsClient() (metrics_client.MetricsClient, func() error, error) {
 		sendUserMetrics,
 		shouldFlushMetricsClientQueueOnEachEvent,
 		do_nothing_metrics_client_callback.NewDoNothingMetricsClientCallback(),
-		logrus_logger_converter.ConvertLogrusLoggerToAnalyticsLogger(logger),
+		analytics_logger.ConvertLogrusLoggerToAnalyticsLogger(logger),
 	)
 
 	if err != nil {
@@ -79,7 +79,7 @@ func GetSegmentClient() (metrics_client.MetricsClient, func() error, error) {
 		sendUserMetrics,
 		shouldFlushMetricsClientQueueOnEachEvent,
 		do_nothing_metrics_client_callback.NewDoNothingMetricsClientCallback(),
-		logrus_logger_converter.ConvertLogrusLoggerToAnalyticsLogger(logger),
+		analytics_logger.ConvertLogrusLoggerToAnalyticsLogger(logger),
 	)
 
 	if err != nil {